requirements.txt
django==2.0
djangorestframework==3.8.2
social-auth-app-django==2.1.0
djangorestframework-jwt
channels==2.1.2
channels_redis
pymongo==3.7.1
mysqlclient